# Conflicts:
#	src/main/java/com/inventoryapp/demo/services/ShopsSalesListingService.java

+++ src/main/java/com/inventoryapp/demo/dtos/ManagerWarehouseCheckinListingDTO.java
@@ -10,11 +10,11 @@ public class ManagerWarehouseCheckinListingDTO {
     private String category;
     private long quantity;
     private String supplierName;
-    private long priceListPerUnit;
-    private long priceSupplierPerUnit;
+    private Double priceListPerUnit;
+    private Double priceSupplierPerUnit;
     private LocalDateTime createDateTime;
 
-    public ManagerWarehouseCheckinListingDTO(String category, long quantity, long priceListPerUnit, long priceSupplierPerUnit, String supplierName, LocalDateTime createDateTime) {
+    public ManagerWarehouseCheckinListingDTO(String category, long quantity, Double priceListPerUnit, Double priceSupplierPerUnit, String supplierName, LocalDateTime createDateTime) {
         this.category = category;
         this.quantity = quantity;
         this.priceListPerUnit = priceListPerUnit;
