+++ src/main/java/com/inventoryapp/demo/dtos/ShopsStockItemDto.java
@@ -7,7 +7,7 @@ public class ShopsStockItemDto {
     private Long position;
     private String shop;
     private String category;
-    private Long quantity;
-    private Long priceSalesPerUnit;
-    private Long priceListPerUnit;
+    private long quantity;
+    private Double priceSalesPerUnit;
+    private Double priceListPerUnit;
 }
