+++ src/main/java/com/inventoryapp/demo/dtos/WarehouseGetAllItemsDTO.java
@@ -9,14 +9,14 @@ import javax.persistence.Id;
 public class WarehouseGetAllItemsDTO {
     private String category;
     private long quantity;
-    private long priceListPerUnit;
+    private Double priceListPerUnit;
     @Id
     @GeneratedValue
     private long id;
     public WarehouseGetAllItemsDTO() {
     }
 
-    public WarehouseGetAllItemsDTO(String category, long quantity, long priceListPerUnit) {
+    public WarehouseGetAllItemsDTO(String category, long quantity, Double priceListPerUnit) {
         this.category = category;
         this.quantity = quantity;
         this.priceListPerUnit = priceListPerUnit;
