+++ src/main/java/com/inventoryapp/demo/dtos/WarehouseSupplierItemDTO.java
@@ -9,10 +9,10 @@ public class WarehouseSupplierItemDTO {
     private String category;
     private long quantity;
     private String supplierName;
-    private long priceListPerUnit;
-    private long priceSupplierPerUnit;
+    private Double priceListPerUnit;
+    private Double priceSupplierPerUnit;
 
-    public WarehouseSupplierItemDTO(String category, long quantity, long priceListPerUnit, long priceSupplierPerUnit, String supplierName) {
+    public WarehouseSupplierItemDTO(String category, long quantity, Double priceListPerUnit, Double priceSupplierPerUnit, String supplierName) {
         this.category = category;
         this.quantity = quantity;
         this.priceListPerUnit = priceListPerUnit;
